Added sentence to Example A
Added a sentence to Example A to clarify that grouping isn't needed when STRING_AGG is the only item in the select. 
I was confused initially when I didn't see a GROUP BY in the first three examples (A, B, and C).  I believe a mention at the end of Example A would have helped me.
Note: new sentence does not indicate that GROUP BY is needed when there is more than one item in the select list, because it isn't required if all the items in the select list are aggregate functions. For example you do not need a GROUP BY if you change Example A to: SELECT Min(FirstName), Max(FirstName), STRING_AGG (FirstName, CHAR(13)) AS csv FROM Person.Person;